+++ backend/internal/configs/config.go
@@ -1,0 +1,13 @@
+package configs
+
+type Config struct {
+	Database Database `yaml:"database"`
+}
+
+type Database struct {
+	Host     string `yaml:"host"`
+	Port     string `yaml:"port"`
+	Username string `yaml:"username"`
+	Password string `yaml:"password"`
+	DBName   string `yaml:"dbname"`
+}