+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat: added config loading
+* 🎉 feat: added config loading
 * 🐛 fix: docker fixed
 * 🐛 fix: workflow fixed
 * 🎉 feat: CD workflow for frontend added and a bug on phone size fixed